// marketing-contentstack-next/types/contentful.d.ts
export interface IBannerFields {
  /** Internal title */
  internalTitle: string;

  /** Text */
  text: Document;

  /** Link text */
  linkText: string;

  /** slug */
  slug: string;

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface IBanner extends Entry<IBannerFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'banner';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IButtonFields {
  /** Internal name */
  internalName: string;

  /** Button text */
  buttonText: string;

  /** Icon */
  icon?: Asset | undefined;

  /** Variant */
  variant: 'primary' | 'secondary' | 'loud';

  /** slug */
  slug: string;
}

export interface IButton extends Entry<IButtonFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'button';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICtaFields {
  /** Internal name */
  internalName: string;

  /** Headline */
  headline: Document;

  /** Subline */
  subline?: Document | undefined;

  /** Buttons */
  buttons?: IButton[] | undefined;

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface ICta extends Entry<ICtaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'cta';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IFeatureFields {
  /** Internal name */
  internalName?: string | undefined;

  /** Headline */
  headline: Document;

  /** Subline */
  subline?: Document | undefined;

  /** Button */
  button?: IButton[] | undefined;

  /** Image */
  image: Asset;

  /** Image position */
  imagePosition?: 'right' | 'left' | undefined;

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface IFeature extends Entry<IFeatureFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'feature';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IFooterFields {
  /** Internal title */
  internalTitle: string;

  /** Footer links */
  footerLinks: IButton[];

  /** Copyright */
  copyright: Document;
}

export interface IFooter extends Entry<IFooterFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'footer';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IFormFields {
  /** Internal Name */
  internalName?: string | undefined;
}

export interface IForm extends Entry<IFormFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'form';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHeroFields {
  /** Internal name */
  internalName?: string | undefined;

  /** Headline */
  headline: Document;

  /** Subline */
  subline?: Document | undefined;

  /** Buttons */
  buttons?: IButton[] | undefined;

  /** Image */
  image: Asset;

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'hero';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHubspotFormFields {
  /** Internal Name */
  internalName: string;

  /** Hubspot Form ID */
  hubspotFormId: string;

  /** Hubspot Portal ID */
  hubspotPortalId: string;
}

/** Connect a Hubspot form using its unique GUID and portal ID. */

export interface IHubspotForm extends Entry<IHubspotFormFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'hubspotForm';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface INavigationFields {
  /** Internal title */
  internalTitle?: string | undefined;

  /** Navigation links */
  navigationLinks: IButton[];

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface INavigation extends Entry<INavigationFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'navigation';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface INtExperienceFields {
  /** Name */
  nt_name: string;

  /** Description */
  nt_description?: string | undefined;

  /** Type */
  nt_type: 'nt_experiment' | 'nt_personalization';

  /** Config */
  nt_config: Record<string, any>;

  /** Audience */
  nt_audience?: INtAudience | undefined;

  /** Variants */
  nt_variants?: Entry<{ [fieldId: string]: unknown }>[] | undefined;
}

/** Ninetailed Experience */

export interface INtExperience extends Entry<INtExperienceFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'nt_experience';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IPricingPlanFields {
  /** Internal name */
  internalName?: string | undefined;

  /** Title */
  title: Document;

  /** Price */
  price: Document;

  /** Frequency */
  frequency?: '/month' | '/week' | '/day' | undefined;

  /** Discounted price */
  discountedPrice?: Document | undefined;

  /** Description */
  description?: Document | undefined;

  /** Button */
  button: IButton;

  /** Most popular */
  mostPopular?: boolean | undefined;
}

export interface IPricingPlan extends Entry<IPricingPlanFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'pricingPlan';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IPricingTableFields {
  /** Internal name */
  internalName?: string | undefined;

  /** Headline */
  headline: Document;

  /** Subline */
  subline: Document;

  /** Pricing plans */
  pricingPlans: IPricingPlan[];

  /** Ninetailed */
  nt_experiences?: INtExperience[] | undefined;
}

export interface IPricingTable extends Entry<IPricingTableFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'pricingTable';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}